pull key iter class back into inner class in LookupData
