+++ src/main/java/com/upserve/uppend/lookup/LookupData.java
@@ -335,9 +335,9 @@ public class LookupData implements AutoCloseable, Flushable {
     }
 
     static Stream<LookupKey> keys(Path path) {
-        LookupKeyIterator iter;
+        KeyIterator iter;
         try {
-            iter = new LookupKeyIterator(path);
+            iter = new KeyIterator(path);
         } catch (IOException e) {
             throw new UncheckedIOException("unable to create key iterator for path: " + path, e);
         }
@@ -347,6 +347,58 @@ public class LookupData implements AutoCloseable, Flushable {
                 Spliterator.DISTINCT | Spliterator.NONNULL | Spliterator.SIZED
         );
         return StreamSupport.stream(spliter, true).onClose(iter::close);
+    }
+
+    private static class KeyIterator implements Iterator<LookupKey>, AutoCloseable {
+        private final Path path;
+        private final Path keysPath;
+        private final FileChannel chan;
+        private final FileChannel keysChan;
+        private final int numKeys;
+        private int keyIndex = 0;
+
+        KeyIterator(Path path) throws IOException {
+            this.path = path;
+            chan = FileChannel.open(path, StandardOpenOption.READ);
+            numKeys = (int) (chan.size() / 16);
+            keysPath = path.resolveSibling("keys");
+            keysChan = numKeys > 0 ? FileChannel.open(keysPath, StandardOpenOption.READ) : null;
+        }
+
+        int getNumKeys() {
+            return numKeys;
+        }
+
+        @Override
+        public boolean hasNext() {
+            return keyIndex < numKeys;
+        }
+
+        @Override
+        public LookupKey next() {
+            LookupKey key;
+            try {
+                key = LookupData.readKey(chan, keysChan, keyIndex);
+            } catch (IOException e) {
+                throw new UncheckedIOException("unable to read at key index " + keyIndex + " from " + path, e);
+            }
+            keyIndex++;
+            return key;
+        }
+
+        @Override
+        public void close() {
+            try {
+                chan.close();
+            } catch (IOException e) {
+                log.error("trouble closing: " + path, e);
+            }
+            try {
+                keysChan.close();
+            } catch (IOException e) {
+                log.error("trouble closing: " + keysPath, e);
+            }
+        }
     }
 
     static int numEntries(FileChannel chan) throws IOException {
